requirements.txt
streamlit==1.35.0
tensorflow==2.16.1
keras==3.3.3
Pillow
numpy
huggingface-hub
streamlit-webrtc
av
